+++ bailian-20231229/alibabacloud_bailian20231229/client.py
@@ -810,12 +810,16 @@ class Client(OpenApiClient):
             query['TableIds'] = request.table_ids_shrink
         if not DaraCore.is_null(request.chunk_mode):
             query['chunkMode'] = request.chunk_mode
+        if not DaraCore.is_null(request.connect_id):
+            query['connectId'] = request.connect_id
         if not DaraCore.is_null(request.database):
             query['database'] = request.database
         if not DaraCore.is_null(request.datasource_code):
             query['datasourceCode'] = request.datasource_code
         if not DaraCore.is_null(request.enable_headers):
             query['enableHeaders'] = request.enable_headers
+        if not DaraCore.is_null(request.knowledge_scene):
+            query['knowledgeScene'] = request.knowledge_scene
         if not DaraCore.is_null(request.meta_extract_columns_shrink):
             query['metaExtractColumns'] = request.meta_extract_columns_shrink
         if not DaraCore.is_null(request.pipeline_commercial_cu):
@@ -911,12 +915,16 @@ class Client(OpenApiClient):
             query['TableIds'] = request.table_ids_shrink
         if not DaraCore.is_null(request.chunk_mode):
             query['chunkMode'] = request.chunk_mode
+        if not DaraCore.is_null(request.connect_id):
+            query['connectId'] = request.connect_id
         if not DaraCore.is_null(request.database):
             query['database'] = request.database
         if not DaraCore.is_null(request.datasource_code):
             query['datasourceCode'] = request.datasource_code
         if not DaraCore.is_null(request.enable_headers):
             query['enableHeaders'] = request.enable_headers
+        if not DaraCore.is_null(request.knowledge_scene):
+            query['knowledgeScene'] = request.knowledge_scene
         if not DaraCore.is_null(request.meta_extract_columns_shrink):
             query['metaExtractColumns'] = request.meta_extract_columns_shrink
         if not DaraCore.is_null(request.pipeline_commercial_cu):